drf_requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.27
django-db-connection-pool==1.2.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
SQLAlchemy==2.0.44
sqlparams==6.2.0
sqlparse==0.5.4
typing_extensions==4.13.2
utils==1.0.2
